converts the main page into a server component
