+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,45 +1,7 @@
-import NextAuth, { NextAuthOptions } from "next-auth";
-import GitHubProvider from "next-auth/providers/github";
 import { NextApiHandler } from "next";
+import NextAuth from "next-auth";
 
-import { PrismaAdapter } from "@next-auth/prisma-adapter";
-import { prisma } from "@/lib/prisma";
-
-const authOptions: NextAuthOptions = {
-  providers: [
-    GitHubProvider({
-      clientId: process.env.GITHUB_CLIENT_ID!,
-      clientSecret: process.env.GITHUB_CLIENT_SECRET!,
-      authorization: {
-        params: {
-          scope: "read:user user:email",
-        },
-      },
-    }),
-  ],
-  session: {
-    strategy: "jwt",
-  },
-  adapter: PrismaAdapter(prisma),
-  callbacks: {
-    async signIn({ user, account, profile }) {
-      if (account?.provider === "github") {
-        user.email = profile?.email ?? user.email;
-      }
-      return true;
-    },
-    async session({ session, user, token }) {
-      session.user = user;
-      return session;
-    },
-    async jwt({ token, user }) {
-      if (user) {
-        token.user = user;
-      }
-      return token;
-    },
-  },
-};
+import { authOptions } from "@/lib/auth";
 
 const handler: NextApiHandler = (req, res) => NextAuth(req, res, authOptions);
